Install utilities on image
Install build utilities we use in the image so we do not have to check
and install our standard build tools in every project that uses them.

Rearrange Dockerfile to optimize for caching.  Still not fully
optimized as some of the more slowly changing layers that are
app-specific are after the installation steps.