addNodeSupport.ts line 127 seems like a strange place to associate the repo that contains the smoke tests and the team that I  want  it to run in. Mainly because the  choice to  tie this goal to sample-sdm  happens in machine.ts line 101, and it is tied to StagingKubernetesDeployGoals which doesn't seem like it should be specific for sample-sdm.
I  don't know  what token  this will get when running, but it needs access to team AHF8B2MBL.

I run the sdm with child_process.spawn (rather than spawnAndWatch) so that I can run the smoke tests on it, and then kill it.There may  be better ways to do this.

I clone the smoke tests with GitCommandGitProject.cloned rather than the  project loader to avoid nesting doWithProject (which seems wierd but may  be fine) and to simplify handling the result. This won't handle smoke tests on BitBucket, so I should probably change this at some point.

I've run executeSmokeTests directly, but did not check in the test because it contained a token. I don't know how that could be committed to the project, the test also does a lot of real I/O and is time consuming. I haven't tested it in the context of a running atomist-sdm (specifically to ensure that StagingKubernetesDeployGoals executes with  this step  properly).

It currently adds a GitHub status to the commit if there is a sha. I can add specific notifications (to the channel associated with the sdm), or the user who made the commit.

These will  currently be run for a new push, but might need to be run continuously to ensure that everything is currently running  properly. It should probably get kicked off when the smoke tests change as well.

When tests fail, there may have been a change in the sdm, a  change in the  smoke tests, neither changed, or both.
If neither changed, and we are running the tests on an interval then we can assume a runtime issue (a service may be down, good info when about to run a demo).
If both changed then we don't really have any info about the cause.
A change to the sdm or the test hints at a root cause, and we could call out the author of  the commit specifically, espcially if the tests consistently fail afterward. If it is still failing for the next commit,  then it would be nice to let that author know that it was already failing (espcially if it failed in the same way).