Exclude *.d.ts from adding license headers
Update the addAtomistHeader exclude pattern to exclude .d.ts files and
remove the exclusion of index.ts and types.ts since the proper fix for
those, and the one that has been implemented, is to generate those
files as part of the build and not include them in source code.  The
pattern that was there no longer worked due to the migration from src
to lib.

Closes #43