Provide autofix for tslint.json
Add autofix to keep tslint.json up to date.  Create new standard
tslint.json that no longer extends tslint:recommended since that can
change, but was created from that set plus our common modifications to
it.  Not sure how we'll keep that up to date as tslint:recommended
changes.  It will be interesting to see what happens when the build
fails due to the autofix.  This build would have, but I corrected it.

Remove the unused smoke test code.

Remove nyc/istanbul test coverage as it was not correct.