Move TSLint from build step to code inspection
Move execution of TSLint to a code inspection that runs after the
build rather than being part of the build.  Move the code inspections
to run after the build so the build and its publication of artifacts
occur regardless of whether the code inspection fails but overall goal
status will be failed.  Add autoCodeInspection as dependency to the
downstream release goals so they are prevented from running should
code inspection fails.  Unfortunately, the buttons for approving the
other dependency goal still appears, although pressing it does not
result in downstream steps executing, as one would wish.  The TSLint
autofix still runs prior to the build.  Also remove TypeDoc generation
from the build.

Move typescript and tslint from devDependencies to dependencies since
we use this project's tslint version and configuration rather than
each projects'.

Use sdm-pack-issue review listener on TSLint auto code fix that
creates and assigns an issue if there are violations.  Provide no more
than 20 comments on review so we do not try to create a GitHub issue
with a body that is too large.  If there are more than 20, truncate to
20 and add an item indicating only the first 20 were included.

Create and use another listener that fails the auto code inspection
goal if there are comments.

Update TSLint configuration, making stylistic and newer rule
violations warnings rather than errors to ease transition to the new
rules for projects that are not up to date.